chore(*): replaced logging with logback, refactored access to objects
Objects should in general not be  public. Always pass the needed objects
to classes you need them in. This will help in separating concerns as it's
easier to notice usage of lots of objects and dependency issues.

DB here is one such thing - this should be refactored so that each using
class is getting an optimized smaller version.